Rename Sync-R1 GRPO bundle and expand setup docs
