+++ sync_r1_grpo/configs/showo_demo.yaml
