+++ sync_r1_grpo/configs/showo_demo_512x512.yaml
